requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0